+++ CHANGES.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [HEAD]
 ### Added
 - Added `Grove.Buzzer.on/1` and `Grove.Buzzer.off/1`.
+- Added `Grove.Relay`.
 
 ## [0.2.0] - 2016-08-15
 ### Added